Get rid deprecated numpy dtypes np.int np.float and np.bool

+++ bgcval2/timeseries/timeseriesTools.py
@@ -227,7 +227,7 @@ def getHorizontalSlice(nc, coords, details, layer, data=''):
     if type(layer) in [
             type(0),
             np.int64,
-            np.int,
+            int,
     ]:
         k = layer
         try:
